Hybrid transport: Switch to event-driven discovery (#98)
## Changes

* Switches from a periodic polling approach, to an event-driven
discovery. It uses btleplug's event system to wait for a BLE
advertisement. This is guaranteed to occur after the QR code is
generated, making this safe.
* Fixes iOS discovery limitation, by removing UUID filter for BlueZ
device discovery. This seems to be a limitation in BlueZ (it [wouldn't
be the first](https://github.com/deviceplug/btleplug/issues/289)), that
I could not find a workaround for. It should have no significant
downside, however, as discovered devices that don't include the relevant
service data are ignored (we do not attempt connecting to them).

## Testing

Tested on an iPhone, and Android 15. Example `webauthn_cable` succeeds
consistently.